feat: add MySQL migrations (#16)
* feat: add driver name

* feat: add mysql support

* chore: move sql package to infrastructure

* chore: rename typemap to sqltype

* chore: add types vars

* fix: some fixes for pg schema loader

* chore: remove some calls prepared type

* chore: remove some calls prepared type

* chore: remove calls prepared type

* style: fix lint

* chore: use MapGoType

* chore: remove unless files

* chore: add json schema format date

* chore: add manager for ddl builders

* feat: update mysql ddl

* chore: move down queries to ddl builder

* chore: move create/drop sequence to postgres ddl builder

* chore: cut ddl builder contract

* chore: simple

* fix: fix diagram template

* fix: fix lint

* fix: fix test

* fix: fix test

* fix: fix test

+++ functest/testdata/pg_go-entities/entities/entities.go
@@ -6,12 +6,12 @@ import (
 )
 
 type Country struct {
-	ID   int64  `db:"id"`
+	ID   int    `db:"id"`
 	Name string `db:"name"`
 }
 
 type User struct {
-	ID          int64           `db:"id"`
+	ID          int             `db:"id"`
 	Name        string          `db:"name"`
 	CountryID   sql.NullInt64   `db:"country_id"`
 	Balance     float64         `db:"balance"`
