+++ functest/testdata/pg_go-entity-repository/internal/entities/country.go
@@ -1,6 +1,6 @@
 package entities
 
 type Country struct {
-	ID   int64  `db:"id"`
+	ID   int    `db:"id"`
 	Name string `db:"name"`
 }
